Restructure the repo into directories
