+++ .dockerignore
@@ -3,5 +3,4 @@
 !package-lock.json
 !index.js
 !schema.js
-!lib/*
-!vendor/*
+!src/*
