+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "start": "node index.js",
     "build": "docker build -t mrorz/cofacts-url-resolver:latest .",
-    "lint": "eslint .",
+    "lint": "eslint src/ tests/",
     "lint:fix": "eslint --fix .",
     "test": "jest"
   },
@@ -34,6 +34,6 @@
     "node": ">=10.0.0"
   },
   "jest": {
-    "setupTestFrameworkScriptFile": "./test-setup.js"
+    "setupTestFrameworkScriptFile": "./tests/setup.js"
   }
 }
