+++ src/lib/catchUnhandledRejection.js
