+++ src/lib/normalize.js
