+++ src/lib/scrap.js
