+++ src/typeDefs/ResolveError.graphql
@@ -1,0 +1,19 @@
+enum ResolveError {
+  # DNS cannot resolve the given URL
+  NAME_NOT_RESOLVED
+
+  # Malformed URL
+  INVALID_URL
+
+  # The target URL cannot be reached at all
+  NOT_REACHABLE
+
+  # File download, etc
+  UNSUPPORTED
+
+  # HTTPS related error, such as invalid certificates
+  HTTPS_ERROR
+
+  UNKNOWN_SCRAP_ERROR
+  UNKNOWN_YOUTUBE_ERROR
+}
