+++ src/typeDefs/UrlResolveResult.graphql
@@ -1,0 +1,42 @@
+type UrlResolveResult {
+  """
+  The exact URL givne as input
+  """
+  url: String
+
+  """
+  Canonical URL extracted from HTML source
+  """
+  canonical: String
+
+  """
+  Page title extracted from HTML source
+  """
+  title: String
+
+  """
+  Summary of the given page. Generated via API or readability.js
+  """
+  summary: String
+
+  """
+  Fetched HTML
+  """
+  html: String
+
+  """
+  URL of the detected top image for the page
+  """
+  topImageUrl: String
+
+  """
+  HTTP status of fetch result. 0 for timeouts, etc.
+  """
+  status: Int
+
+  """
+  Known fatal error during the resolution process. These errors are not related to HTTP, thus
+  cannot be represented as HTTP status codes.
+  """
+  error: ResolveError
+}
